+++ plugins/diagramEditor/languages/jetbrains.mps.lang.editor.diagram/source_gen/jetbrains/mps/lang/editor/diagram/editor/FigureParameterMappingName_Actions.java
@@ -9,6 +9,7 @@ import jetbrains.mps.openapi.editor.cells.CellActionType;
 import jetbrains.mps.editor.runtime.cells.AbstractCellAction;
 import jetbrains.mps.lang.smodel.generator.smodelAdapter.SPropertyOperations;
 import jetbrains.mps.smodel.adapter.structure.MetaAdapterFactory;
+import jetbrains.mps.editor.runtime.highlight.DeletionApproverUtil;
 import jetbrains.mps.lang.smodel.generator.smodelAdapter.SNodeOperations;
 
 public class FigureParameterMappingName_Actions {
@@ -28,6 +29,9 @@ public class FigureParameterMappingName_Actions {
       if (SPropertyOperations.getString(node, MetaAdapterFactory.getProperty(0xceab519525ea4f22L, 0x9b92103b95ca8c0cL, 0x110396eaaa4L, 0x110396ec041L, "name")) != null) {
         SPropertyOperations.set(node, MetaAdapterFactory.getProperty(0xceab519525ea4f22L, 0x9b92103b95ca8c0cL, 0x110396eaaa4L, 0x110396ec041L, "name"), null);
       } else {
+        if (DeletionApproverUtil.approve(editorContext, node)) {
+          return;
+        }
         SNodeOperations.deleteNode(node);
       }
     }
@@ -44,6 +48,9 @@ public class FigureParameterMappingName_Actions {
       if (SPropertyOperations.getString(node, MetaAdapterFactory.getProperty(0xceab519525ea4f22L, 0x9b92103b95ca8c0cL, 0x110396eaaa4L, 0x110396ec041L, "name")) != null) {
         SPropertyOperations.set(node, MetaAdapterFactory.getProperty(0xceab519525ea4f22L, 0x9b92103b95ca8c0cL, 0x110396eaaa4L, 0x110396ec041L, "name"), null);
       } else {
+        if (DeletionApproverUtil.approve(editorContext, node)) {
+          return;
+        }
         SNodeOperations.deleteNode(node);
       }
     }
